Remove any Python SDK roots found in project roots (PY-5459)
It's quite common to add the project root to sys.path in order to make
it available in other Python projects at run-time. Unfortunately, it
breaks excluded directories settings since indexing SDK roots in the
IntelliJ platform ignores excluded directories since they are
project-level, not application-level.

Since for most setups having a directory in both SDK roots and project
(content and source) roots is redundant, we now delete it from the SDK
roots. The user can add it back manually via interpreters settings in
PyCharm and via module dependencies in IntelliJ if they need this entry
in their SDK roots.